Moved fixture from chunking test to conftest - expect to use in different parts of the code.
