+++ tests/snappy_pipeline/workflows/test_chunk.py
@@ -3,12 +3,8 @@
 """
 
 from collections import defaultdict
-import io
 from operator import attrgetter
-import random
-import textwrap
 
-from biomedsheets.io_tsv import read_germline_tsv_sheet
 from biomedsheets.shortcuts import GermlineCaseSheet, is_background, is_not_background
 import pytest
 
@@ -38,468 +34,6 @@ def build_pedigree_size_dictionary(all_pedigrees):
         pedigree_count_per_index_dict[index_secondary_id] = size
     # Return
     return pedigree_count_per_index_dict
-
-
-def random_gender():
-    """Random gender.
-    :return: Returns either 'F' or 'M' by change.
-    """
-    genders = ["F", "M"]
-    # pseudo-random generator is unsafe for crypto use, but OK here
-    return random.choice(genders)  # nosec
-
-
-def random_affected_status():
-    """Random affected status.
-    :return: Returns either 'Y' or 'N' by change.
-    """
-    affected = ["Y", "N"]
-    # pseudo-random generator is unsafe for crypto use, but OK here
-    return random.choice(affected)  # nosec
-
-
-def set_trio_ids(index_i):
-    """Define trio identifiers.
-
-    :param index_i: Index identifier. Example if '1', final identifier will be 'P001'.
-    :type index_i: int
-
-    :return: Returns tuple with trio identifiers (index, mother, father).
-    Example: ('P001', 'P002', 'P003').
-    """
-    # Initialise variable
-    sampleid_pattern = "P{i}"
-    # Set sample ids for trio
-    index_sampleid = sampleid_pattern.format(i=str(index_i).zfill(3))
-    mother_i = index_i + 1
-    mother_sampleid = sampleid_pattern.format(i=str(mother_i).zfill(3))
-    father_i = index_i + 2
-    father_sampleid = sampleid_pattern.format(i=str(father_i).zfill(3))
-    # Return
-    return index_sampleid, mother_sampleid, father_sampleid
-
-
-def get_entry_for_sample_sheet(index_i, entry_type="trio"):
-    """Sample sheet entry text.
-
-    :param index_i: Index id. Example if '1', final id will be 'P001'
-    :type index_i: int
-
-    :param entry_type: Type of entry: 'solo', 'duo', 'trio'. Default: 'trio'.
-    :type entry_type: str, optional
-
-    :return: Returns text used to define a trio case in sample sheet.
-    """
-    # Initialise variables
-    entry_pattern = (
-        "{sampleid}\t{father_sampleid}\t{mother_sampleid}\t{gender}\t{affected}"
-        "\tWGS\tAgilent SureSelect Human All Exon V6\t{sampleid}\t.\n"
-    )
-
-    # Validate entry option
-    valid_options = ["solo", "duo", "trio"]
-    if entry_type not in valid_options:
-        valid_options_str = ", ".join(valid_options)
-        err_msg = "Options '{in_}' is not valid. Valid entry types options: {valid}".format(
-            in_=entry_type, valid=valid_options_str
-        )
-        raise ValueError(err_msg)
-
-    # Set sample ids for trio
-    index_sampleid, mother_sampleid, father_sampleid = set_trio_ids(index_i=index_i)
-
-    # Set entries for trio
-    index_gender = random_gender()
-    index_entry = entry_pattern.format(
-        sampleid=index_sampleid,
-        father_sampleid=father_sampleid,
-        mother_sampleid=mother_sampleid,
-        gender=index_gender,
-        affected="Y",
-    )
-    mother_affected = random_affected_status()
-    mother_entry = entry_pattern.format(
-        sampleid=mother_sampleid,
-        father_sampleid=".",
-        mother_sampleid=".",
-        gender="F",
-        affected=mother_affected,
-    )
-    father_affected = random_affected_status()
-    father_entry = entry_pattern.format(
-        sampleid=father_sampleid,
-        father_sampleid=".",
-        mother_sampleid=".",
-        gender="M",
-        affected=father_affected,
-    )
-
-    # Return accordingly
-    if entry_type == "solo":
-        index_entry = index_entry.replace(mother_sampleid, ".").replace(father_sampleid, ".")
-        return index_entry
-    elif entry_type == "duo":
-        index_entry = index_entry.replace(father_sampleid, ".")
-        return index_entry + mother_entry
-    elif entry_type == "trio":
-        return index_entry + mother_entry + father_entry
-    return None
-
-
-def get_entry_for_sample_sheet_with_custom_fields(index_i, batch_number, entry_type="trio"):
-    """Sample sheet entry text with custom fields
-
-    :param index_i: Index id. Example if '1', final id will be 'P001'
-    :type index_i: int
-
-    :param batch_number: Batch number.
-    :type batch_number: int
-
-    :param entry_type: Type of entry: 'solo', 'duo', 'trio'. Default: 'trio'.
-    :type entry_type: str, optional
-
-    :return: Returns text used to define a trio case in sample sheet with custom fields 'familyId'
-    and 'batchNo'.
-    """
-    # Initialise variables
-    entry_pattern = (
-        "{familyid}\t{sampleid}\t{father_sampleid}\t{mother_sampleid}\t{gender}\t{affected}"
-        "\t{batch}\tWGS\tAgilent SureSelect Human All Exon V6\t{sampleid}\t.\n"
-    )
-
-    # Validate entry option
-    valid_options = ["solo", "duo", "trio"]
-    if entry_type not in valid_options:
-        valid_options_str = ", ".join(valid_options)
-        err_msg = "Options '{in_}' is not valid. Valid entry types options: {valid}".format(
-            in_=entry_type, valid=valid_options_str
-        )
-        raise ValueError(err_msg)
-
-    # Set sample ids for trio
-    index_sampleid, mother_sampleid, father_sampleid = set_trio_ids(index_i=index_i)
-
-    # Set family id
-    familyid = "FAM_" + index_sampleid
-
-    # Set entries for trio
-    index_gender = random_gender()
-    index_entry = entry_pattern.format(
-        familyid=familyid,
-        sampleid=index_sampleid,
-        father_sampleid=father_sampleid,
-        mother_sampleid=mother_sampleid,
-        gender=index_gender,
-        affected="Y",
-        batch=batch_number,
-    )
-    mother_affected = random_affected_status()
-    mother_entry = entry_pattern.format(
-        familyid=familyid,
-        sampleid=mother_sampleid,
-        father_sampleid=".",
-        mother_sampleid=".",
-        gender="F",
-        affected=mother_affected,
-        batch=batch_number,
-    )
-    father_affected = random_affected_status()
-    father_entry = entry_pattern.format(
-        familyid=familyid,
-        sampleid=father_sampleid,
-        father_sampleid=".",
-        mother_sampleid=".",
-        gender="M",
-        affected=father_affected,
-        batch=batch_number,
-    )
-
-    # Return accordingly
-    if entry_type == "solo":
-        index_entry = index_entry.replace(mother_sampleid, ".").replace(father_sampleid, ".")
-        return index_entry
-    elif entry_type == "duo":
-        index_entry = index_entry.replace(father_sampleid, ".")
-        return index_entry + mother_entry
-    elif entry_type == "trio":
-        return index_entry + mother_entry + father_entry
-    return None
-
-
-@pytest.fixture
-def header_germline_sheet():
-    """Returns germline TSV file header with wildcard '{entries}'"""
-    return textwrap.dedent(
-        """
-        [Custom Fields]
-        key\tannotatedEntity\tdocs\ttype\tminimum\tmaximum\tunit\tchoices\tpattern
-        libraryKit\tngsLibrary\tEnrichment kit\tstring\t.\t.\t.\t.\t.
-
-        [Data]
-        patientName\tfatherName\tmotherName\tsex\tisAffected\tlibraryType\tlibraryKit\tfolderName\thpoTerms
-        {entries}
-        """
-    ).lstrip()
-
-
-@pytest.fixture
-def header_germline_sheet_with_custom_fields():
-    """Returns germline TSV file header with custom fields and with wildcard '{entries}'"""
-    return textwrap.dedent(
-        """
-        [Custom Fields]
-        key\tannotatedEntity\tdocs\ttype\tminimum\tmaximum\tunit\tchoices\tpattern
-        batchNo\tbioEntity\tBatch No.\tinteger\t.\t.\t.\t.\t.
-        familyId\tbioEntity\tFamily\tstring\t.\t.\t.\t.\t.
-        libraryKit\tngsLibrary\tEnrichment kit\tstring\t.\t.\t.\t.\t.
-
-        [Data]
-        familyId\tpatientName\tfatherName\tmotherName\tsex\tisAffected\tbatchNo\tlibraryType\tlibraryKit\tfolderName\thpoTerms
-        {entries}
-        """
-    ).lstrip()
-
-
-@pytest.fixture
-def text_large_cohort_trios_only():
-    """Defines arbitrary large cohort entries for sample sheet - trio cases only"""
-    # Initialise variables
-    n_patients_cohorts = 501  # 167 indexes
-    full_cohort_text = ""
-    # Create cohort
-    index_i = 1
-    while index_i < n_patients_cohorts:
-        new_trio = get_entry_for_sample_sheet(index_i=index_i)
-        # Append
-        full_cohort_text = full_cohort_text + new_trio
-        # Update index counter
-        index_i += 3
-    # Return
-    # Remove last newline so biomedsheet doesn't interpret it as an empty entry
-    return full_cohort_text.rstrip("\n")
-
-
-@pytest.fixture
-def text_large_cohort_diverse():
-    """Defines arbitrary large cohort entries for sample sheet - diverse cases"""
-    # Initialise variables
-    full_cohort_text = ""
-    index_i = 1
-    # Create cohort trios - 30 cases, 90 samples
-    for _ in range(30):
-        new_trio = get_entry_for_sample_sheet(index_i=index_i)
-        # Append
-        full_cohort_text = full_cohort_text + new_trio
-        # Update index counter
-        index_i += 3
-    # Create cohort duo - 30 cases, 60 samples
-    for _ in range(30):
-        new_duo = get_entry_for_sample_sheet(index_i=index_i, entry_type="duo")
-        # Append
-        full_cohort_text = full_cohort_text + new_duo
-        # Update index counter
-        index_i += 2
-    # Create cohort solo - 50 cases, 50 samples
-    for _ in range(50):
-        new_solo = get_entry_for_sample_sheet(index_i=index_i, entry_type="solo")
-        # Append
-        full_cohort_text = full_cohort_text + new_solo
-        # Update index counter
-        index_i += 1
-    # Return
-    # Remove last newline so biomedsheet doesn't interpret it as an empty entry
-    return full_cohort_text.rstrip("\n")
-
-
-@pytest.fixture
-def text_medium_cohort_solo_only():
-    """Defines arbitrary medium cohort entries for sample sheet - solo case only"""
-    # Initialise variables
-    full_cohort_text = ""
-    index_i = 900
-    # Create cohort solo - 99
-    for _ in range(99):
-        new_solo = get_entry_for_sample_sheet(index_i=index_i, entry_type="solo")
-        # Append
-        full_cohort_text = full_cohort_text + new_solo
-        # Update index counter
-        index_i += 1
-    # Return
-    # Remove last newline so biomedsheet doesn't interpret it as an empty entry
-    return full_cohort_text.rstrip("\n")
-
-
-@pytest.fixture
-def text_medium_cohort_diverse_with_custom_fields():
-    """Defines arbitrary medium cohort entries for sample sheet with custom fields -
-    diverse cases"""
-    # Initialise variables
-    full_cohort_text = ""
-
-    # Create index list
-    indexes_list = list(range(1, 31, 3))  # 10 trio cases, 30 samples
-    indexes_list += list(range(31, 51, 2))  # 10 duo cases, 20 samples
-    indexes_list += list(range(51, 101, 1))  # 50 solo cases, 50 samples
-
-    # Create case type list - reflects `indexes_list` structure
-    case_list = 10 * ["trio"]
-    case_list += 10 * ["duo"]
-    case_list += 50 * ["solo"]
-
-    # Create batch list
-    batch_list = []
-    for i in range(1, 8, 1):  # 7 batches, 10 cases in each
-        batch_list.extend(10 * [i])
-    batch_list = batch_list[::-1]  # reverse order -> [start:stop:step]
-
-    # Create cohort
-    for index, batch, case in zip(indexes_list, batch_list, case_list):
-        case = get_entry_for_sample_sheet_with_custom_fields(
-            index_i=index, batch_number=batch, entry_type=case
-        )
-        # Append
-        full_cohort_text = full_cohort_text + case
-
-    # Return
-    # Remove last newline so biomedsheet doesn't interpret it as an empty entry
-    return full_cohort_text.rstrip("\n")
-
-
-@pytest.fixture
-def text_small_cohort_trio_with_custom_fields():
-    """Defines small cohort entries for sample sheet with custom fields - two trio cases"""
-    return (
-        textwrap.dedent(
-            """
-        FAM_P001\tP001\tP002\tP003\tF\tY\t1\tWGS\tAgilent SureSelect Human All Exon V6\tP001\t.
-        FAM_P001\tP002\t.\t.\tM\tN\t1\tWGS\tAgilent SureSelect Human All Exon V6\tP002\t.
-        FAM_P001\tP003\t.\t.\tF\tN\t3\tWGS\tAgilent SureSelect Human All Exon V6\tP003\t.
-        FAM_P004\tP004\tP005\tP006\tM\tY\t2\tWGS\tAgilent SureSelect Human All Exon V6\tP004\t.
-        FAM_P004\tP005\t.\t.\tM\tN\t2\tWGS\tAgilent SureSelect Human All Exon V6\tP005\t.
-        FAM_P004\tP006\t.\t.\tF\tY\t2\tWGS\tAgilent SureSelect Human All Exon V6\tP006\t.
-        """
-        )
-        .lstrip()
-        .strip()
-    )
-
-
-@pytest.fixture
-def tsv_large_cohort_trios_only_germline_sheet(header_germline_sheet, text_large_cohort_trios_only):
-    """Returns contents for large cohort germline TSV file - trio cases only"""
-    return header_germline_sheet.format(entries=text_large_cohort_trios_only)
-
-
-@pytest.fixture
-def tsv_large_cohort_diverse_cases_germline_sheet(header_germline_sheet, text_large_cohort_diverse):
-    """Returns contents for large cohort germline TSV file - diverse cases"""
-    return header_germline_sheet.format(entries=text_large_cohort_diverse)
-
-
-@pytest.fixture
-def tsv_medium_cohort_solo_only(header_germline_sheet, text_medium_cohort_solo_only):
-    """Returns contents for medium cohort germline TSV file - solo cases only"""
-    return header_germline_sheet.format(entries=text_medium_cohort_solo_only)
-
-
-@pytest.fixture
-def tsv_medium_cohort_diverse_ngs_kit_solo_only(
-    header_germline_sheet, text_medium_cohort_solo_only
-):
-    """Returns contents for medium cohort germline TSV file - solo cases only with two different
-    NGS sequencing kits.
-    """
-    # Change sequencing kit for 1/2 the entries
-    entries_list = text_medium_cohort_solo_only.splitlines()
-    for i in range(50):
-        entries_list[i] = entries_list[i].replace(
-            "Agilent SureSelect Human All Exon V6", "Illumina TruSeq PCR-free"
-        )
-    new_entries_test = "\n".join(entries_list)
-    return header_germline_sheet.format(entries=new_entries_test)
-
-
-@pytest.fixture
-def tsv_medium_cohort_diverse_with_custom_features(
-    header_germline_sheet_with_custom_fields, text_medium_cohort_diverse_with_custom_fields
-):
-    """Returns contents for medium cohort germline with custom fields TSV file - diverse cases"""
-    return header_germline_sheet_with_custom_fields.format(
-        entries=text_medium_cohort_diverse_with_custom_fields
-    )
-
-
-@pytest.fixture
-def tsv_small_cohort_with_custom_features(
-    header_germline_sheet_with_custom_fields, text_small_cohort_trio_with_custom_fields
-):
-    """Returns contents for small cohort germline with custom fields TSV file - two trio cases"""
-    return header_germline_sheet_with_custom_fields.format(
-        entries=text_small_cohort_trio_with_custom_fields
-    )
-
-
-@pytest.fixture
-def germline_sample_sheet_object_large_cohort_trios_only(
-    tsv_large_cohort_trios_only_germline_sheet,
-):
-    """Returns GermlineCaseSheet object with large cohort - trio cases only"""
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_large_cohort_trios_only_germline_sheet)
-    return GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
-
-
-@pytest.fixture
-def germline_sample_sheet_object_large_cohort_diverse_cases(
-    tsv_large_cohort_diverse_cases_germline_sheet,
-):
-    """Returns GermlineCaseSheet object with large cohort - trio cases onl."""
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_large_cohort_diverse_cases_germline_sheet)
-    return GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
-
-
-@pytest.fixture
-def germline_sample_sheet_object_medium_cohort_solo_cases_background(tsv_medium_cohort_solo_only):
-    """Returns GermlineCaseSheet object with medium background cohort - solo cases only"""
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_medium_cohort_solo_only)
-    sheet = GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
-    sheet.sheet.json_data["extraInfoDefs"]["is_background"] = {"type": "boolean", "default": False}
-    sheet.sheet.extra_infos["is_background"] = True
-    return sheet
-
-
-@pytest.fixture
-def germline_sample_sheet_object_medium_cohort_solo_only_diverse_kits(
-    tsv_medium_cohort_diverse_ngs_kit_solo_only,
-):
-    """Returns GermlineCaseSheet object with medium cohort with two different sequencing kits -
-    only solo cases
-    """
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_medium_cohort_diverse_ngs_kit_solo_only)
-    return GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
-
-
-@pytest.fixture
-def germline_sample_sheet_object_medium_cohort_diverse_with_custom_features(
-    tsv_medium_cohort_diverse_with_custom_features,
-):
-    """Returns GermlineCaseSheet object with medium cohort with custom features - diverse cases"""
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_medium_cohort_diverse_with_custom_features)
-    return GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
-
-
-@pytest.fixture
-def germline_sample_sheet_object_small_cohort_with_custom_features(
-    tsv_small_cohort_with_custom_features,
-):
-    """Returns GermlineCaseSheet object with medium cohort with custom features - diverse cases"""
-    # Create dna sample sheet based on germline sheet
-    germline_sheet_io = io.StringIO(tsv_small_cohort_with_custom_features)
-    return GermlineCaseSheet(sheet=read_germline_tsv_sheet(germline_sheet_io))
 
 
 @pytest.fixture
